+++ llm4s.g8/src/main/g8/default.properties
@@ -1,6 +1,7 @@
-name=llm4s-demo
-organization=com.example
+name=llm4s-template
+package=org.llm4s.template
 version=0.1.0
 llm4s_version=0.10.0
 scala_version=3.3.1
+munit_version=0.7.29
 directory=\$name\$
