+++ llm4s.g8/src/main/g8/project/plugins.sbt
@@ -1,1 +1,3 @@
-addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.5.2")
+addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.7.5")
+addSbtPlugin("org.scalameta" % "sbt-munit" % "0.7.29")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.2.5") // if using coverage
