+++ llm4s.g8/src/main/g8/src/main/scala/$package__packaged$/PromptExecutor.scala
@@ -1,7 +1,11 @@
-import org.llm4s.openai.OpenAI
-import org.llm4s.trace.Tracing
+package $package;format="package"$
 
+/**
+ * This code is part of the Giter8 template llm4s.g8 in llm4s project, which provides a set standard template/archetype
+ * for improve developer onboarding, creating new projects using the llm4s library.
+ */
 object PromptExecutor {
+
   def run(prompt: String): Unit = {
     val tracer = Tracing.create()
     tracer.traceEvent(s"Executing prompt: " + prompt)
@@ -17,12 +21,5 @@ object PromptExecutor {
     }
 
     tracer.traceEvent("Prompt execution finished")
-  }
-}
-
-object Main {
-  def main(args: Array[String]): Unit = {
-    val prompt = args.headOption.getOrElse("Explain what a Monad is in Scala")
-    PromptExecutor.run(prompt)
   }
 }
