--- llm4s.g8/src/main/g8/src/test/scala/MainSpec.scala
@@ -1,7 +1,0 @@
-package test
-
-class MainSpec extends munit.FunSuite {
-  test("basic assertion") {
-    assert(1 + 1 == 2)
-  }
-}
